+++ data/appData/api/src/main/java/kanti/todoer/data/appdata/AppDataRepository.kt
@@ -4,11 +4,11 @@ import kotlinx.coroutines.flow.Flow
 
 interface AppDataRepository {
 
-	val currentPlanId: Flow<Int>
+	val currentPlanId: Flow<Long>
 
-	suspend fun setCurrentPlan(planId: Int)
+	suspend fun setCurrentPlan(planId: Long)
 
 	suspend fun deleteCurrentPlan()
 
-	suspend fun deleteIfCurrent(planId: Int)
+	suspend fun deleteIfCurrent(planId: Long)
 }
